testdrive: add coverage of avro ocf files
