+++ src/testdrive/lib.rs
@@ -16,6 +16,7 @@ use self::error::{Error, InputError, ResultExt};
 use self::parser::LineReader;
 
 mod action;
+mod avro;
 mod parser;
 mod protobuf;
